# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Software delivery store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  version: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  downloadUrl: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  version: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
  imageUrl?: string;
  /** @minLength 1 */
  downloadUrl: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  version?: string;
  /** @minimum 0 */
  price?: number;
  category?: string;
  imageUrl?: string;
  downloadUrl?: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  productId: number;
  productName: string;
  productPrice: number;
  licenseKey: string;
  status: string;
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerEmail: string;
  productId: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  revenue: number;
}

export interface StoreStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
  ordersByCategory: CategoryCount[];
}

export type ListProductsParams = {
category?: string;
};

export type LookupOrdersParams = {
email: string;
};