# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Software delivery store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Software product operations
  - name: orders
    description: Order operations
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order and receive a license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed with license key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/lookup:
    get:
      operationId: lookupOrders
      tags: [orders]
      summary: Look up orders by customer email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders for that email
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          description: Missing email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details with license key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Product:
      type: object
      required: [id, name, description, version, price, category, imageUrl, downloadUrl, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        version:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        downloadUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    ProductInput:
      type: object
      required: [name, description, version, price, category, downloadUrl]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        category:
          type: string
          minLength: 1
        imageUrl:
          type: string
        downloadUrl:
          type: string
          minLength: 1

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
        imageUrl:
          type: string
        downloadUrl:
          type: string

    Order:
      type: object
      required: [id, customerName, customerEmail, productId, productName, productPrice, licenseKey, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        productId:
          type: integer
        productName:
          type: string
        productPrice:
          type: number
        licenseKey:
          type: string
        status:
          type: string
        createdAt:
          type: string

    OrderInput:
      type: object
      required: [customerName, customerEmail, productId]
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          minLength: 1
        productId:
          type: integer

    StoreStats:
      type: object
      required: [totalProducts, totalOrders, totalRevenue, recentOrders, ordersByCategory]
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        ordersByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"

    CategoryCount:
      type: object
      required: [category, count, revenue]
      properties:
        category:
          type: string
        count:
          type: integer
        revenue:
          type: number