# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Software delivery store API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  productId: number;
  productName: string;
  productPrice: number;
  licenseKey: string;
  status: string;
  createdAt: string;
}