# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Software delivery store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  version: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  downloadUrl: string;
  createdAt: string;
  updatedAt: string;
}