# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Software delivery store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  version: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
  imageUrl?: string;
  /** @minLength 1 */
  downloadUrl: string;
}