# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Software delivery store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  version?: string;
  /** @minimum 0 */
  price?: number;
  category?: string;
  imageUrl?: string;
  downloadUrl?: string;
}